+++ test/selector-validation.js
@@ -234,4 +234,10 @@ describe('selector validation', function() {
       }
     }
   );
+
+  describe('with @keyframes rule', function() {
+    it('does not complain about keyframe selectors', function() {
+      assertSuccess('/** @define Foo */ @keyframes fade { 0% { opacity: 0; } 100% { opacity: 1; } }');
+    });
+  });
 });