+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+=== Head
+
+* Ensure that `@keyframes` selectors are always skipped.
+
 === 2.3.0 (November 24, 2015)
 
 * Add ability to ignore custom properties via `postcss-bem-linter: ignore` comments and the `ignoreCustomProperties` pattern.